+++ packages/api-common/source/rcp/index.ts
@@ -1,2 +1,2 @@
 export { Processor } from "./processor.js";
-export * as Utils from "./utils.js";
+export * as Utils from "./utilities.js";
