+++ packages/contracts/source/index.ts
@@ -5,4 +5,4 @@ export * as Contracts from "./contracts/index.js";
 export * as Events from "./events.js";
 export * as Exceptions from "./exceptions/index.js";
 export * from "./identifiers.js";
-export * as Utils from "./utils.js";
+export * as Utils from "./utilities.js";
