+++ packages/crypto-transaction/source/validation/index.ts
@@ -1,4 +1,4 @@
 export * from "./formats.js";
 export * from "./keywords.js";
 export * from "./schemas.js";
-export { extendSchema } from "./utils.js";
+export { extendSchema } from "./utilities.js";
