+++ packages/crypto-transaction/source/validation/schemas.test.ts
@@ -11,7 +11,7 @@ import cryptoJson from "../../../core/bin/config/testnet/core/crypto.json";
 import { describe, Sandbox } from "../../../test-framework/source";
 import { makeKeywords } from "./keywords";
 import { schemas, transactionBaseSchema } from "./schemas";
-import { extendSchema, signedSchema, strictSchema } from "./utils";
+import { extendSchema, signedSchema, strictSchema } from "./utilities";
 
 describe<{
 	sandbox: Sandbox;
