+++ packages/p2p/source/hapi-nes/utilities.ts
