+++ packages/transaction-pool-service/source/index.ts
@@ -7,4 +7,4 @@ export * from "./sender-state.js";
 export * from "./service.js";
 export * from "./service-provider.js";
 export * from "./storage.js";
-export * from "./utils.js";
+export * from "./utilities.js";
